moved flags back to bools

+++ h5tools/create_fiber_index_mask.py
@@ -59,13 +59,13 @@ gal_flag = FibIndex.get_gal_flag()
 meteor_flag = FibIndex.get_meteor_flag()
 badfib_flag = FibIndex.get_badfiber_flag()
 
-FibIndex.fiber_table["flag_shot"] = shot_flag.astype(int)
-FibIndex.fiber_table["flag_throughput"] = throughput_flag.astype(int)
-FibIndex.fiber_table["flag_badamp"] = amp_flag.astype(int)
-FibIndex.fiber_table["flag_largegal"] = gal_flag.astype(int)
-FibIndex.fiber_table["flag_meteor"] = meteor_flag.astype(int)
-FibIndex.fiber_table["flag_satellite"] = sat_flag.astype(int)
-FibIndex.fiber_table["flag_badfib"] = badfib_flag.astype(int)
+FibIndex.fiber_table["flag_shot"] = shot_flag
+FibIndex.fiber_table["flag_throughput"] = throughput_flag
+FibIndex.fiber_table["flag_badamp"] = amp_flag
+FibIndex.fiber_table["flag_largegal"] = gal_flag
+FibIndex.fiber_table["flag_meteor"] = meteor_flag
+FibIndex.fiber_table["flag_satellite"] = sat_flag
+FibIndex.fiber_table["flag_badfib"] = badfib_flag
 
 FibIndex.fiber_table["flag"] = (
     amp_flag
@@ -75,7 +75,7 @@ FibIndex.fiber_table["flag"] = (
     * throughput_flag
     * badfib_flag
     * sat_flag
-).astype(int)
+)
 
 flag_table = FibIndex.fiber_table[
     "fiber_id",
